Added initial version of validate.py
